Changed: Poll components folder structure
